Move IGitCommands members from GitCommandsInstance to GitModule
